Fix Ruby 3.4-dev compatibility
Ref: https://bugs.ruby-lang.org/issues/19117
Ref: https://bugs.ruby-lang.org/issues/16495

Now backtrace use constisten `'` quotes rather than a mix of backtick and quote.

Also the label include the class or module names.

But in most case dealing with `Backtrace::Location` objects is
much more performant and clear than to match a regexp on
the full rendered frame. So aside from Ruby compatibility,
this is just better code.