Fix dynamic constant handling
Rather than specifying whether to *skip* the logic, we should
say whether to replace the constant node with a dynamic const assignment.

Then, remove unnecessary type check and move this logic to
the top of the method so we short-circuit any unnecessary
code in that case.

Run two more tests that are only failing because of error differences

Remove `desugar/regex` from failing tests

Remove `desugar/range` from excluded tests

Co-Authored-By: Stan Lo <stan.lo@shopify.com>