Refactor `translateHash` into `translateKeyValuePairs`
The logic of `translateHash` was kind of confusing because it was handling
both vanilla hashes and also assocs/assoc splat nodes. By isolating one
part of that method, which generates key/value pairs, and then calling
that method in both cases, the two code paths become cleaner and more
obvious.