Fix parentheses with non-statement-node body in Prism
Prism parses `(10)` as a `ParenthesesNode` with a `StatementsNode` body
but in `def (10).test; end` it's a `ParehthesesNode` with an
`IntegerNode` directly in the body. This commit ensures we don't cast
to `pm_statements_node` in this case.

We've also added an extra test case based on the example in the Prism
codebase [1], but it doesn't yet pass due to another issue.

[1]: https://github.com/ruby/prism/blob/b3d83f6400a760501a9a360afda8f68d7a4cb805/config.yml#L4072