Fix double block parameter in Prism parser mode
The Prism parser mode was causing a crash with the following code:

```ruby
def foo(&, ...)
end
```

This is invalid Ruby because the forwarding parameter `...` already
includes block forwarding. However, Prism doesn't remove the extra block
param the same way the default parser does. This commit fixes the issue
by skipping adding the block parameter when a forwarding parameter is
present.