Disable prism/legacy pipeline comparison for certain tests
These are cases where we know the resuling desugar trees will differ,
but we're happy with the differences. For example, in some cases, Prism
can throw away less of tree for invalid Ruby code. Previously, we were
doing the comparison in the `verify_prism_regression_tests.sh` script,
but now we're doing it as part of the test suite. As such, a magic
comment seemed like the best way to go.