Handle trailing type aliases comments
This commit ensures we properly process trailing type aliases comments
when walking the body of a node. For example:

```rb
module Foo
  BAR = 123
  #: type bar = Integer
end
```

It extracts a helper method to avoid duplicating the logic for processing
trailing comments.

It also fixes the same issue when the body of the node is empty:

```rb
module Foo
  #: type bar = Integer
end
```

This case was not previously handled in the original RBS rewriter
either. As such, I fixed it there too, also pulling our a helper method
to avoid duplicating the logic.