Use static names for anonymous parameters
The original RBS rewriter runs after the parser has assigned unique
names to anonymous parameters, which it can then pull out of the method
definition. With the Prism RBS rewriter, we can't generate unique names
until later on when the translator is run. We can use static names
instead as `*`, `**`, and `&` are guaranteed to be unique anyway (in
fact this is already the case for the block param).