Fix missing predicate in case match
```ruby
case
in x
end
```

This is invalid Ruby because the predicate is missing. The original
parser interprets the invalid code differently and ends up with a
`RaiseUnimplemented` node, but in our case we can use `EmptyTree`
to handle this instead, since it's not an unimplemented feature, it's
invalid syntax.